Added sentence citation relation
